+++ src/mon-util.h
@@ -25,6 +25,7 @@
 const char *describe_race_flag(int flag);
 void create_mon_flag_mask(bitflag *f, ...);
 struct monster_race *lookup_monster(const char *name);
+bool monster_is_aquatic(const struct monster_race *race);
 struct monster_base *lookup_monster_base(const char *name);
 bool match_monster_bases(const struct monster_base *base, ...);
 void update_mon(struct monster *mon, struct chunk *c, bool full);